Review Changes
Made changes requested in code review.
