+++ src/App.tsx
@@ -8,11 +8,11 @@ import {
 } from 'react-router-dom';
 //Remember to import tsx files for pages!
 //Import components that are inside pages in page tsx files themselves
-import NextButton from './components/nextButton';
-import Sidebar from './components/sidebar';
-import Landing from './pages/landing';
-import Python from './pages/python';
-import Variable from './pages/variable';
+import Landing from './components/pages/landing';
+import Python from './components/pages/python';
+import Variable from './components/pages/variable';
+import NextButton from './components/shared/nextButton';
+import Sidebar from './components/shared/sidebar';
 
 function App(): JSX.Element {
   return (
