--- src/components/booleanDropdown.tsx
@@ -1,14 +1,0 @@
-import Dropdown from 'react-dropdown';
-import '../styles/booleanDropdown.scss';
-
-const options = ['True', 'False'];
-
-interface Props {
-  change: () => void;
-}
-
-function BooleanDropdown(props:Props):JSX.Element  {
-  return <Dropdown className='drop' options={options} value='False' arrowClassName='arrow' onChange={() => props.change()}/>;
-}
-
-export default BooleanDropdown;
