+++ src/components/pages/landing.tsx
@@ -1,4 +1,4 @@
-import '../styles/landing.scss';
+import '../../styles/sidebar.scss';
 function Landing(): JSX.Element {
   return (
     <div>
