+++ src/components/pages/variable.tsx
@@ -1,4 +1,4 @@
-import '../styles/variable.scss';
+import '../../styles/variable.scss';
 function Variable(): JSX.Element {
   return (
     <div>
