+++ src/styles/python.scss
@@ -13,15 +13,15 @@
   width: 450px;
 }
 
-#w {
+.white {
   color: white;
 }
 
-#o {
+.orange {
   color: $main-orange;
 }
 
-#b {
+.blue {
   color: $main-blue;
 }
